+++ docs/changelog.md
@@ -4,6 +4,7 @@
 * Remove postcss-nested package as it's not needed after WordPress version 6.7.
 * Convert wp cli inner block generator's useBlockProps to useInnerBlockProps.
 * Have aquamin_get_post_content() parse blocks more accurately, and support "raw" option (e.g. for default_content filter usage).
+* Prevent reordering of less than 2 items with EntitySelector.
 * Fixed EntitySelector block UI component flexibility update.
 
 ## 7.3.0 - May 22 2025