+++ docs/changelog.md
@@ -6,6 +6,7 @@
 * Have aquamin_get_post_content() parse blocks more accurately, and support "raw" option (e.g. for default_content filter usage).
 * Prevent reordering of less than 2 items with EntitySelector.
 * Fixed EntitySelector block UI component flexibility update.
+* Default to showing appender when using useInnerBlocksProps(useBlockProps()) (see Gutenberg GitHub issue #70275).
 
 ## 7.3.0 - May 22 2025
 * Fixed JavaScript linting errors.