+++ docs/changelog.md
@@ -8,6 +8,7 @@
 * Fixed EntitySelector block UI component flexibility update.
 * Default to showing appender when using useInnerBlocksProps(useBlockProps()) (see Gutenberg GitHub issue #70275).
 * Created a classNameUI() utility for adding simple class names to blocks.
+* Created getFauxInnerBlocks() utility for exposing inner blocks to dynamic blocks in the editor.
 
 ## 7.3.0 - May 22 2025
 * Fixed JavaScript linting errors.